+++ src/tests/address_reputation_test.ts
@@ -1,6 +1,6 @@
 import { AddressReputation } from "../coinbase/address_reputation";
 
-describe("AddressReputation", () => {
+describe("AddressReputation for risky address", () => {
   let addressReputation: AddressReputation;
 
   beforeEach(() => {
@@ -40,13 +40,51 @@ describe("AddressReputation", () => {
     });
   });
 
-  it("returns the risky as true", () => {
-    expect(addressReputation.risky).toBeTruthy();
+  it("returns the risky as true for score < 0", () => {
+    expect(addressReputation.risky).toBe(true);
   });
 
   it("returns the string representation of the address reputation", () => {
     expect(addressReputation.toString()).toBe(
       "AddressReputation(score: -90, metadata: {unique_days_active: 1, total_transactions: 1, token_swaps_performed: 1, bridge_transactions_performed: 1, smart_contract_deployments: 1, longest_active_streak: 1, lend_borrow_stake_transactions: 1, ens_contract_interactions: 1, current_active_streak: 1, activity_period_days: 1})",
     );
+  });
+
+  it("should return risky as false for a score > 0", () => {
+    addressReputation = new AddressReputation({
+      score: 90,
+      metadata: {
+        unique_days_active: 1,
+        total_transactions: 1,
+        token_swaps_performed: 1,
+        bridge_transactions_performed: 1,
+        smart_contract_deployments: 1,
+        longest_active_streak: 1,
+        lend_borrow_stake_transactions: 1,
+        ens_contract_interactions: 1,
+        current_active_streak: 1,
+        activity_period_days: 1,
+      },
+    });
+    expect(addressReputation.risky).toBe(false);
+  });
+
+  it("should return risky as false for a score=0", () => {
+    addressReputation = new AddressReputation({
+      score: 0,
+      metadata: {
+        unique_days_active: 1,
+        total_transactions: 1,
+        token_swaps_performed: 1,
+        bridge_transactions_performed: 1,
+        smart_contract_deployments: 1,
+        longest_active_streak: 1,
+        lend_borrow_stake_transactions: 1,
+        ens_contract_interactions: 1,
+        current_active_streak: 1,
+        activity_period_days: 1,
+      },
+    });
+    expect(addressReputation.risky).toBe(false);
   });
 });